+++ tests/versionNineFieldParser.test.ts
@@ -84,6 +84,10 @@ ZTZTAN`;
     expect(parser.parseCountry()).toBe('United States');
   });
 
+  it('should parse country code correctly', () => {
+    expect(parser.parseCountryCode()).toBe('USA');
+  });
+
   it('should parse document ID correctly', () => {
     expect(parser.parseString('documentId')).toBe('00121300011116256231');
   });